Added a fake cooling function to the dummy environment. Fixed a bug in scenarios with no unsatisfied OptimalConditions. Changed method signatures to use concrete types for performance improvements. Work in progress.

+++ SmartNode/Logic/Models/MapekModels/SimulationTick.cs
@@ -6,7 +6,7 @@ namespace Logic.Models.MapekModels
     {
         public required int TickIndex { get; init; }
 
-        public required double TickDurationSeconds { get; init; }
+        public required int TickDurationSeconds { get; init; }
 
         public required IEnumerable<ActuationAction> ActionsToExecute { get; init; }
     }
